requirements.txt
pandas
enum
streamlit
random
matplotlib.pyplot
plotly.express
plotly.graph
time